fix tests, tidy packages
