--- src/main/java/com/hotels/hcommon/hive/metastore/MetaStoreClientException.java
@@ -1,30 +1,0 @@
-/**
- * Copyright (C) 2018 Expedia Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.hotels.hcommon.hive.metastore;
-
-
-public class MetaStoreClientException extends RuntimeException {
-  private static final long serialVersionUID = 1L;
-
-  public MetaStoreClientException(String message) {
-    super(message);
-  }
-
-  public MetaStoreClientException(String message, Throwable cause) {
-    super(message, cause);
-  }
-
-}
