+++ src/main/java/com/hotels/hcommon/hive/metastore/client/supplier/HiveMetaStoreClientSupplier.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.hotels.hcommon.hive.metastore.client;
+package com.hotels.hcommon.hive.metastore.client.supplier;
 
 import org.apache.hadoop.hive.conf.HiveConf;
