+++ src/test/java/com/hotels/hcommon/hive/metastore/client/closeable/CloseableMetaStoreClientInvocationHandlerTest.java
@@ -39,7 +39,6 @@ public class CloseableMetaStoreClientInvocationHandlerTest {
   private @Mock IMetaStoreClient client;
   private @Mock HiveMetaStoreClientCompatibility compatibility;
 
-
   @Test
   public void typical() throws Throwable {
     invocationHandler = new CloseableMetaStoreClientInvocationHandler(client, compatibility);
