+++ src/test/java/com/hotels/hcommon/hive/metastore/client/supplier/HiveMetaStoreClientSupplierTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.hotels.hcommon.hive.metastore.client;
+package com.hotels.hcommon.hive.metastore.client.supplier;
 
 import static org.mockito.Matchers.eq;
 import static org.mockito.Mockito.verify;
